backend/requirements.txt
asgiref==3.5.0
django-filter==21.1
Django==4.0.4
djangorestframework==3.13.1
pip==22.0.4
python-dotenv==0.20.0
pytz==2022.1
setuptools==57.0.0
sqlparse==0.4.2
tzdata==2022.1
wheel==0.36.2
gunicorn==20.1.0
djangorestframework-simplejwt==5.1.0
djoser==2.1.0
django-extra-fields==3.0.2
Pillow==9.1.0
psycopg2-binary==2.9.3
